Fix of the method SetWidth() that was taking an int as argument, instead of a double as it should be
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@29802 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ topcaf/dataobjects/include/TOPCAFDigit.h
@@ -88,7 +88,7 @@ namespace Belle2 {
     //! TDC bin
     void SetTDCBin(double tdc_bin) {m_tdc_bin = tdc_bin;}
     //! Width
-    void SetWidth(int width) {m_width = width;}
+    void SetWidth(double width) {m_width = width;}
     //! amplitude
     void SetAmplitude(double amp) {m_amp = amp;}
     //! sampling rate
